relise 1.0

+++ app/Http/Controllers/ContactsController.php
@@ -2,6 +2,8 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Http\Request;
 use App\Contact;
 use App\Telephone;
 
@@ -13,7 +15,7 @@ class ContactsController extends Controller {
     }
 
 	public function index() {
-		$contacts = Contact::all();
+		$contacts = Contact::all()->where('ctc_uti_id_ce', '=', Auth::id());
         $updateContact = new Contact();
         $actionR = '/contact/add';
         $contactCategories = Contact::getValeursEnum('contacts', 'ctc_categorie');
@@ -22,30 +24,30 @@ class ContactsController extends Controller {
             compact( 'contacts', 'contactCategories', 'telephoneCategories', 'updateContact', 'actionR') );
 	}
 
-//	public function show( Contact $contact ) { //Contact::find(wildcard);
-//
-//		return view( 'contact.un', compact( 'contact' ) );
-//	}
-
 	public function delete( Contact $contact ) {
-        $contact->deleteContact();
+        if ($contact->ctc_uti_id_ce === Auth::id()) {
+            $contact->deleteContact();
+        }
 		return redirect( '/contact' );
 	}
 
 	public function edit( Contact $contact ) {
-        $contacts = Contact::all();
-        $updateContact = $contact;
+        if ($contact->ctc_uti_id_ce === Auth::id()) {
+            $contacts = Contact::all()->where('ctc_uti_id_ce', '=', Auth::id());
+            $updateContact = $contact;
 //        $actionR = "{{ route('edit.contact'), ['contact' => $contact->ctc_id] }}";
-        $actionR = "/contact/update/" . $contact->ctc_id;
-        $contactCategories = Contact::getValeursEnum('contacts', 'ctc_categorie');
-        $telephoneCategories = Contact::getValeursEnum('telephones', 'tel_type');
-
+            $actionR = "/contact/update/" . $contact->ctc_id;
+            $contactCategories = Contact::getValeursEnum('contacts', 'ctc_categorie');
+            $telephoneCategories = Contact::getValeursEnum('telephones', 'tel_type');
+        }
         return view( 'contact.tout',
             compact( 'contacts', 'contactCategories', 'telephoneCategories', 'updateContact', 'actionR') );
     }
 
-	public function update( Contact $contact ) {
-        $contact->updateContact();
+	public function update( Contact $contact, Request $request ) {
+        if ($contact->ctc_uti_id_ce === Auth::id()) {
+            $contact->updateContact($this->getContactFormData($request));
+        }
 
         return redirect( '/contact' );
     }
@@ -56,23 +58,22 @@ class ContactsController extends Controller {
      * @return redirection sur la page des contacts
      */
 
-    public function add(Contact $contact) {
+    public function add(Contact $contact, Request $request) {
+            $contact->addContact($this->getContactFormData($request));
 
-        $contact->addContact($this->getContactFormData());
-
-        return redirect("/contact");
+            return redirect("/contact");
     }
 
-    private function getContactFormData(){
+    private function getContactFormData($request){
         return [
-            'ctc_prenom' => request('prenom'),
-            'ctc_nom' =>  request('nom'),
-            'ctc_categorie' => request('contact-category'),
-            'ctc_uti_id_ce' => 1
+            'ctc_prenom' => $request->prenom,
+            'ctc_nom' =>  $request->nom,
+            'ctc_categorie' => $request->contact_category,
+            'ctc_uti_id_ce' => Auth::id()
         ];
     }
 
-    private function getTelephonesFormData(){
-
-    }
+//    private function getTelephonesFormData(){
+//
+//    }
 }
